Move the trivial Default* types to the same files as the interface they implement, so that alternative built-in implementations can be see together.
