--- src/Fixie/DefaultExecution.cs
@@ -1,10 +1,0 @@
-﻿namespace Fixie;
-
-public sealed class DefaultExecution : IExecution
-{
-    public async Task Run(TestSuite testSuite)
-    {
-        foreach (var test in testSuite.Tests)
-            await test.Run();
-    }
-}
