+++ src/Fixie/IExecution.cs
@@ -7,3 +7,12 @@ public interface IExecution
     /// </summary>
     Task Run(TestSuite testSuite);
 }
+
+public sealed class DefaultExecution : IExecution
+{
+    public async Task Run(TestSuite testSuite)
+    {
+        foreach (var test in testSuite.Tests)
+            await test.Run();
+    }
+}
